Reuse marker-symbol for all counters
Instead of having additional properties, introduce "magic" marker names
that simplify overall design and remove ambiguity.

Now, to show the mark as a number, name it "-number".  For multiple counters,
append the name of the group, e.g.  "-number-museums".  The same works with
the letters, e.g.  "-letter" or "-letter-museums".

Change-Id: Ib62ce349c0856d8b821f77330d874508497e8e65