Fix preview failure because of missing map data.
When previewing a brand new article (has not been saved yet),
Kartographer may load group data but the API won't return
any data (the article does not exist in the database yet).

This patch adds a safe check to avoid failing, but in theory
Kartographer should not try to load any map data that we know does
not exist.

Bug: T135546
Change-Id: Id8248d4d507d50b6700255dbd1791e0b05cff1a8