Add front-end documentation inline and generate public docs with jsduck.
This is a huge effort for documenting the front-end code.
The current architecture may not be perfect as it is not totally stable yet,
so be patient if you find code samples that are not working.
This is a big step forward that will help each contributor understand, review,
stabilize and maintain the codebase, particularly the public APIs.

Bug: T140203
Change-Id: I62da008a1aa50e424081b629b507304d90a1ce08