Have the sidebar adjust to interface language direction
The map is inside the content area, so it adheres to content
direction - which means that if the user views a map in a LTR
wiki but with an RTL interface, the sidebar is not properly
aligned.

Adding the 'direction: ltr;' rule to the LESS file allows it to
go through CSSJanus and override content direction for the map
with the interface direction that the user chose.

Change-Id: Ib0a31b2b52468427647177b66ac6b6c8280a9883