Use require() instead of mw.loader.require()
mw.loader.require() is an internal method for debugging only.
It was a failure on my part to document it as such (until recently).

When T108655 added support for `module.exports` and `require()` to
ResourceLoader, it also ensured the require() function is accessible
to all module execution contexts, similar to how jQuery and mw are
ensured via the implement() closure from load.php, and in debug mode.

Change-Id: I9aae3b6660b4bc6345c24029a56036f669c54228