Fix corruption of title/description wikitext to HTML when editing in VE
The raw GeoJSON editor at the bottom of the VE dialog uses the original
GeoJSON from the wiki page, but what the map renders is the sanitized
version from the server. As part of this sanitization process, the
server parses wikitext to HTML, so that the client can render it.

When an edit is made using the visual tools on the map, this modifies
the internal model of the map, which we then export back to GeoJSON and
load into the raw GeoJSON editor. But now we've replaced wikitext labels
with HTML labels, which will then be dirty-diffed on the wiki page.

To fix this, have the sanitizer back up the original values of the title
and description properties as _origtitle and _origdescription, and then
restore those values when exporting the modified GeoJSON from the map
editor back into the raw GeoJSON editor (and from there into the wiki
page).

Bug: T187230
Change-Id: I3bf36c93491c1f24f2c36dc4885dd8e296d4709c