Simplify State::$data to always be an associative array
This is the only functional change split from the bigger patch
I4b9c5b3, which is otherwise about documentation only.

I figured this is the most obvious, most convenient solution. Inside
the State class an associative array is most easy to work with. For
the relevant JSON encoding (teh only consumer of the data array) it
does not matter if an object or an associative array are returned.
Both result in the same JSON anyway.

Change-Id: Ife664468dd348942a20e129388ef0462665b99b9