Update map language in VE rendering when it changes
Right now there's no way to change the language in the UI, but if you
process a transaction that updates the lang attribute, the CE rendering
doesn't update.

This adds a setLang method to the map to change the language, and calls
it in ce.MWMapsNode. Also adds a convenience method to dm.MWMapsNode to
get the language code out.

Bug: T193659
Change-Id: Ied7c81ff2d2c03e23c5f652fde5803d5aa0bd2e9