Make sure footer is created once
There is a case where the 'getSetupProcess' of the dialog is
opened twice, and because it is asynchronous, the isFirstTimeOpen
is not registered correctly.

Make sure that the process in 'addFooterButton' that creates and
appends the footer elements is only creating and appending new
DOM elements if they don't already exist.

Bug: T194803
Change-Id: I92fd0f3ef12d95c3569d80e8f63737429179fa92