Add OpenStreetMap cycle map to list of Kartographer external services
This change should add the OpenStreetMap cycle map to the "topographic" list
of external mapping services. While I have not discussed this change,
OpenStreetMap.org's default style is already linked and OSM is a "featured"
service; ACME Mapper is currently the only map listed in the "topographic"
section; and ACME Mapper is not an official service and is a web application
built on top of maps run by others. Furthermore, although the OSM map is
primarily a cycle map, ACME Mapper's topographic map does not have coverage
outside the United States, whereas the OSM map has worldwide coverage.

Bug: T196854
Change-Id: I33a4b7ae63a472fe04cbb20f5cb129e46d3eea13