2 PackageFiles conversion fixes
Dialogs has an undefined dependency on mapbox (to extend the
fullscreen button). This became visible after conversion, because the
execution order slightly changed.
A good reproduction is to have a <maplink> on a page, but NO
<mapframe>. When you choose fullscreen, it execs the fullscreen widget
code before loading kartographer.box, which was including mapbox
already.

Additionally, the enablePreview module was not being executed
causing it to not show the coordinate popup on right-clicks in preview
mode.

Follow-up: Ia01026864f11d876d2a7ad365a6403f0a26ea9ea
Change-Id: I308fa7d4e7b088c45a9295cead05621575a68116