Correct autozoom/center calculations on uncollapsed maps
When uncollapsing or opening a section, a map placed there that
requires autozoom, auto fitting to contents, would be offset or show
as the full world map.

This is because such maps rely on bounds calculation, but this fails
when the map size is 0x0 due to being collapsed. These incorrect
values were than stored as the initialmap position.

Make it so that when we uncollapse maps that were never visible, we
reinitialize the view with the correct original position so that
autozoom is able to be rerun.

Also add support for MobileFrontend collapsed sections. Unfortunately
there didn't seem to be a way to make initialy 'uncollapsed' MF
sections work, too many resizes are happening at the same time, it
becomes a mess.

Bug: T192250
Change-Id: I9a04527a536e7ea0e44a8a7558d4c057a8fcb95d