Use requestAnimationFrame to fix livepreview problem
When the map gets added in a wikipage.content hook, it might be
initialized before the node is visible in the dom. This makes it
initialize with the incorrect dimensions and part of the map will not
be visible. This was especially the case with live preview in the
wikitext editor and previously with VE (T125263).

By using requestAnimationFrame we delay until next render cycle.

Bug: T151524
Change-Id: Id305a2a2093e230596c7623821ff2724252563ea