Ensure the map is initialized if external data fails
The map should never be left in an uninitialized state, as the Leaflet
code is not really suited for that. In one case, a single dragging of
the invisible map surface, would lead to an endless loop of logged
errors. This is undesirable, webcalls should be expected to fail from
time to time.

Bug: T257872
Change-Id: I45c20eb143d876c2b5f6365a1dae1dee649fa9d6