Expose first JSON validation error to the user
Without this it's not possible to fix an error in the JSON without
guesswork or using another validation service. A proper formatter
that displays all errors might be better. But this here is the most
minimal improvement I could come up with. It allows the user to fix
one error at a time.

Bug: T223049
Change-Id: I7fd11728fdf1d48636fe75ad050ee7261a2cd262