Simplify handling of layer.dataGroup attribute
All code related to this ….dataGroup property was implemented
in I2545a22 (T151929).

Note how all 3 places store different (!) things in the same
property. One place stores a self-reference (!), one place
stores a full object, and one place looks like it stores the
group ID string.

The only code that ever uses this property is an
`if ( obj.layer.dataGroup )` in ControlLayers.js.

I assume the property is meant to either store a boolean "does
this layer represent a dataGroup, yes or no?" This is how it is
used. Or it is meant to store the group ID the layer belongs
to as some kind of "back-reference". This is what I do in this
patch. Line #472 stores the layer in the group, and line #473
stores the "back-reference" back from the layer to the group
(by ID, not as an actual reference to not have circular
references, these are bad for the garbage collection).

Change-Id: Ieb38ce9e8ba9aa2a015c53c25ac12539077fb9f4