Fix incomplete SimpleStyleParser::normalize() implementation
Compare with SimpleStyleParser::sanitize() just a few lines
above (line #177). sanitize() accepts both arrays of objects
as well as a single object. normalize() accepts only arrays.

This is called from JCMapDataContent::getSafeData() from the
JsonConfig extension. Because of the incomplete implementation
here there is a hack in JsonConfig that forces the input to be
an array. But this hack is bogus as well. It turns arrays of
objects into 2-dimenstional arrays. This is fine for sanitize()
but not for normalize().

With this change it doesn't matter if it's an array and how deep
it's nested.

I checked existing real-world data on Wikimedia Commons and
there are indeed Data:….map pages where the "data" element is an
array of objects, and others where it's a single object.

Bug: T281224
Change-Id: I20dfe70dfd00b564f0e4757b7772f5804f4c2c81