links and staticframes don't have parentMaps
This was throwing errors when closing the fullscreen dialog with the
escape key. The point here was to reset the map position of the
non-fullscreen map to the values of the fullscreen map.

But renderNewMap is used for maplinks and for static mapframes, where
resetting the position was only intended for dynamic maps.

Bug: T267466
Change-Id: I42f5fd59fdb073259518aee8ab052a1c33fee068