Use $content from hook instead of $('.mw-body') for maplink
This allows to execute live preview more than once with attached
maplinks.

The maplinks in indicators are covered by $('.mw-body') but not covered
by hook `wikipage.content`. Therefor the new hook `wikipage.indicators`
is needed.

The support for maplinks in indicators requires the new hook
`wikipage.indicators` introduced in MediaWiki 1.36.

Bug: T280363
Depends-On: I154ebbf049fbce1e872548651953adb1419b6cba
Change-Id: Ia5e7386fa641677a684da4d7466ae160ad89620e