Replace unexpected counters object with array
I find the usage of an stdClass object quite unexpected. All it does
is tracking some numbers, indexed by marker id. We typically use
int[] arrays for this. As far as I can tell the only reason it is an
object is because objects are passed by reference. We can do this
with arrays as well.

Change-Id: If7358869e889a68deacdadc52902d3c2ce814f39