Support versioned live maps
When parsed mapdata isn't already present on the page, fetch it from
the mapdata API, making sure to match the currently viewed page
revision.

Can be disabled with a new, temporary feature flag
`wgKartographerVersionedLiveMaps`, which when set to `false` falls
back to the legacy behavior, of loading mapdata for the the latest
revision of the page.

Bug: T293844
Change-Id: Ibbfb70ea66ac9c72250036c7174c2783b7a825dd