Guard non-critical code agains unexpected user input
What this code does is updating marker symbols from e.g. "foo-counter"
to "foo-1", "foo-2" and so on. This is pointless when the input is not
even valid. There is no good reason why this particular code should
fail in such a case. Validation happens somewhere else.

This patch also replaces property_exists() with isset(). The main
difference is that property_exists() will be true when the property
exists, but is null. This is never useful in these cases.

Bug: T288671
Change-Id: I9abaaf8ed94d7f6f1cf1bf683a1923d15cf98b33