Make SimpleStyleParser a little more robust
Technically there is a validation step in place that _should_ make
sure all objects have this "type" property. However. This is not
obvious just from looking at the code in this patch here. The contrary.
This particular line of code looks like it should expect the property
to be missing, and just not do anything then.

Bug: T304657
Change-Id: I25c2c1c3f4c54eb414bc736108dbd599ac087a85