Workaround for FlaggedRevs stable-revision caching
Dips into FlaggedRevs internals, to determine when we can should pull
from its stable revision cache.  This prevents an extra page parse
and some wasted RevisionOutputCache when viewing static mapframe
thumbnails on a stable revision for an article with drafts.

This is meant to be a temporary workaround and the patch includes
instructions on how to roll back when the upstream issues are
resolved.

Includes instrumentation to log in case our assumptions were
incorrect and this endpoint can be reached when the cache is cold.

Bug: T293905
Bug: T304813
Change-Id: I58af8d7322759a572883d94c13f66d411d9b015b