Refresh MediaWiki globals when loading mapdata
The mw.config constants were being accessed at module load time,
which could be in an entirely different pageview than run time.  This
was causing outdated revision IDs to leak, and resulted in a failure
to render the interactive map immediately after page save.  Wrapping
the instantiation in a function causes the config to be reloaded each
time the method is used.

Another mystery not addressed here is that wgKartographerLiveData is
empty immediately after page save—if this were populated then the
mapdata request would be unnecessary.

Bug: T307650
Change-Id: I577d8cc559d266ed7f71b5ee77634210300d8bd9