Include zoom factor into de-bouncing calculation
The most trivial yet sufficient algorithm we could come up with. The
radius is simply rounded to the 2 most significant digits. The center
point is rounded to a specific number of decimal places, ranging from
4 to 0. We never go higher than 4 (corresponds to ~11m) and never
lower than 0.

The mapping between zoom factor and decimal places is a little
arbitrary, based on experiments with typical screen sizes. Might be
tweaked later, if needed.

Bug: T308235
Change-Id: I295b1fc35499306967cd393f1615b5059b3bdbf8