Add custom styling for nearby markers
When the geojson is converted to the object structure we can use
he created marker objects for custom styling. This is kind of a
POC. The color is the same as in the Wikivoyage nearby feature.

Note that Kartotherian adds the marker color as part of the marker's
properties. But it seems that there's a conversion missing here.

Also note: This does not solve the issue with the mapbox.com
defaults for the simple styled icons. We just avoid that issue.

Bug: T314925
Change-Id: Idfb04befaa167ff3e4d6d428872563dbbeb33dfe