Use more memory efficient hash function
Using the maximum of 36 is a common trick to store a number in a
string that's as short as possible. This does have another advantage
as well: Now points are also considered equal when they are really
close to each other (less than 0.7m). This is intentional. Even if two
points represent two different things, the two markers are not usable
when they sit on top of each other. Since the nearby feature shows some
arbitrary set of articles anyway it should be fine to skip some that
are less relevant.

Bug: T315547
Change-Id: I7c6874f095f57729309425518e61076770ddd719