Show all GeoJSON validation errors
… instead of only the first one, which is often the least relevant
one. This is a direct follow up to what I started in I7fd1172.

The way this JSON schema validation works is that the validator tries
a lot of ways to understand the input. When it fails to figure out a
way that works it will report all of them as failures. That's why we
see a long list of messages. I tried to explain this in the first
line that's shown to the user.

Bug: T223049
Bug: T304748
Bug: T316213
Change-Id: If8611680906047cce9b61a6e613c064aa2a34ae6