Remove not needed XML header from new nearby SVG icons
It's often a good idea to have this header to maximize compatibility.
But it's not needed when the SVG is embedded in CSS. It just makes
the CSS bigger for everyone, even if it goes unused in most sessions.

While we don't have any guideline for this, similar .svg files in
many other codebases do this as well.

Bug: T315861
Change-Id: Ia1b313d92c0ae2c26450e8e93fb6a317b60e5b18