Drop foreign nearby layers earlier
This feature was added via I539d075. The basic idea is to start
dropping nearby points when the zoom level changes to much and the
previously loaded points would either collapse into a dense cluster
in the middle of the screen or all be way outside of the visible
area.

This patch makes the feature feel a little more responsive. The old
layer is dropped much faster → then new nearby data is requested via
the API → then new nearby points appear. This is relevant because it
decreases the time the user might click on one of the old points just
to see it disappear a second later.

Bug: T315547
Change-Id: I0f09249e080d4acd6d68301d39e4ddf60432fe1c