Consider thumbnail width/height when rendering nearby popups
Looks like we just forgot this. The information is already there in
the API response.

This does have a significant user-facing effect. Before, a popup was
shown in a small size as if there is image. A few milliseconds later
the popup size changed. This is not only a visible "FOUC", it also
triggers unnecessary panning in case the popup opens outside of the
visible viewport. Before the map was panned 2 times in the worst
case. Now this happens only 1 time.

Bug: T321305
Bug: T304078
Change-Id: I76846027bda185c61343754d026d81564b9f08b2