Don't do expensive wikitext parsing of empty strings
I'm told the parser might do something meaningful even when the input
is empty. But in this context we really don't care. The only place
where these strings will show up is in tiny popup windows on a map.
In this context I think empty should really mean just that: nothing
to see here.

Bug: T315913
Change-Id: I34802717b76d25a3086d76c372bdae448861144c