Fix error reporting mistaken array of GeoJSON as array of errors
This is a little design flaw. We use the value in the status object
to carry either the validated GeoJSON (more precisely an array of
GeoJSON objects) or an array of validation errors. This is a quick
patch to cover the mistake. A proper fix will follow in a later patch.

Bug: T317873
Bug: T319077
Change-Id: I962ef270001d8fb6bb3e106a7569909fd3746a1a