Show an outline on the map when using keyboard nav
2nd attempt including code to prevent showing the focus when using
the mouse. See Id1dce7181937fee1ba26a07f61ac889f327b421d

As soon as the mouse is used to move the map, the focus frame is
hidden. It will just be visible again when using keys to navigate.

We also want to make sure the focus stays on the map so that the
user can switch between mouse and keyboard navigation as needed.

Bug: T315997
Change-Id: Ib5ae052a9a0c91b00dee8be25b698172fe98abaa