Make getScaleLatLng() calculation more robust
The only way this piece of code can create unexpected results is when
the method is called with an invalid zoom like 0.5 or smaller or
1e+31 or bigger. Let's try if this get's rid of the log spam.

Note this code is not critical at all. All it does is to round the
coordinates displayed in the sidebar to a sensible number of
fractional digits. Any zoom factor of 1 or smaller shows the entire
world and should be rounded to an integer.

Bug: T321603
Change-Id: Ie8ee2d1a8c44e35452f934306658d13ddb30f9f5