Fix uncatched exception when handling non-numeric user input
Both the wikitext as well as the input fields in the VisualEditor
dialog can contain bogus, non-numeric stuff. As of now this causes
exceptions that show up in the JS console and possibly even break
the dialog.

Note there was already a string to number conversion done via the +
in front of the three properties. But it was incomplete, resulting
in e.g. NaN.

Change-Id: Ieba2343504167232226ffe71d3ac4f3ccfc451c0