nearby: Replace hand-crafted debounce with OO.ui.debounce()
Note: I don't fully understand how the OOUI modules depend on each
other. Listing OOUI core shouldn't be necessary. But better do it to
be safe. The debounce function is in OOUI core.

We could also use mw.util.debounce() here. This function is part of
the "mediawiki.util" module. Both should be avaliable already. The
implementation of the two is the same.

Note: It's probably a good idea to reduce the 500ms to something
like 200ms to make it feel "snappier". But this is intentionally not
part of this patch.

Change-Id: I2d14bf4b01662e287209a303441f59085c12d627