Use flex layout for VisualEditor maps dialog
The old behavior was not nice on medium size and smaller screens.
The map on the top was forced to use 70% of the available space. This
ignored entirely how much space the form elements in the lower half
actually need. On my screen the lower half always had a scrollbar for
no good reason.

This is now more relevant because we added form elements, e.g. the
"frameless" toggle.

The proposed behavior is that the lower half is allowed to use as
much space as it needs, but never more than half of the available
space.

Bug: T318834
Change-Id: I712936578db8b9be2250d4e266c77af5862361c7