Fix confusing order of validation error messages
The errors now follow how the elements are ordered in the syntax:
1. The attributes in the XML-style <mapframe …> tag.
2. Groups are also attributes. This feature set is exclusive to
   Wikivoyage and comes second because of this.
3. The GeoJSON content in the tag.

Furthermore we hide the long and possibly confusing JSON schema
validation error list when the <mapframe> tag is not valid anyway.
This means users are encouraged to fix errors in this order. The
second group of errors will only be shown when the first group is
fine. This is rather common in wikitext, e.g. the TemplateData
validation works exactly like this. Here it's especially relevant
because the JSON errors are so complex.

Change-Id: Id783d44423b73ed6925cfe1010e2ec7f12d827ee