Breaking upgrade: mapdata
The 0.8 version of mapdata returns an array of mapdata rather than a object
keyed by group ID.  These IDs were only used as a key for
deduplicating groups.

Mapdata was extracting ExternalData out of groups and the synthesized
ID based on the entire geojson contents is now unnecessary.

Attributions are rendered in this extension rather than by kartotherian.

Bug: T322616
Change-Id: I48b1bc44d9f3d4655cbbd54a08dbfa4267e18e7f