Introduce tiny ParserFunctionTracker service
All we want to do is to add a tracking category or (in future code)
count expensive parser functions. To do this we are forced to pass the
entire parser around even if the code doesn't need to parse anything.
This patch introduces the smalles possible interface that exposes only
what's actually needed. This encapsulates things and makes mocking and
testing easier.

This is done in preparation for T323113 where we want to count
expensive parser functions.

Bug: T323113
Change-Id: I70522b4405896d6e88d6bffa146b4867650c7803