Skip invalid ExternalData without a url
While this is not strictly necessary because of the validation that
happens before this is called, it feels better to have this basic check
for the "url" property added. It makes this service more independent
from where and under which circumstances it is used.

This also reduces the test setup to the most minimal one needed to
prove this part of the code works. Some duplicate test cases are
removed.

Bug: T323113
Change-Id: I5fd8a8feb616e930dcb3de07ffb0660cd130ce53