Replace hard-coded clip rectangle hack with more flexible CSS
Main changes:
* Replace deprecated clip with clip-rect. This makes it possible to
  use relative % values, which is needed to be able to move it to the
  .css file.
* Don't clip when tabbing with the keyboard. This will make the large
  rectangle visible. But this is much better than a clipped rectangle.
  Can be improved later, if this is even possible.

This also fixes the issue that the old focus-rectangle hack was
not active on wikis with $wgKartographerStaticMapframe = true. See
I111497d where we discovered this.

Note this change removes support for IE11. I would argue this is
acceptable for a tiny optimization like this. IE11 users can still
do everything. They just get the original, larger click region below
markers as it was before T141750.

Bug: T326530
Change-Id: I794d692959df21811aeead65978f670c956ac1d9