Fix large focus rectangle on map markers with modern CSS
Alternative to I83d821d as well as I794d692. Tested with Chrome and
Firefox.

I keep the original clip-path hack for two reasons:
1. It does have better browser support.
2. The *= selector is expensive. Combining it with :focus means it
   will only be applied to the one element that is currently focused.

The way this is implemented means that nothing bad will happen on
browsers that don't support this, except that the focus rectangle
will be as big as before. This is only an inconvenience.

Bug: T326530
Change-Id: I256cef4d9bbb9ab5dbd61cf26802d1b7330267a7