Use more robust object-fit: cover for marker icons
"none" doesn't stretch the image in any way but maps it 1:1 to the
pixels on the screen. This might be inappropriate when the image is
e.g. double resolution (@2x). In this case the <img> tag might be
35x90 but the .png file is 70x180.

Examples:
https://maps-beta.wmflabs.org/v4/marker/pin-l-theatre+000.png
vs.
https://maps-beta.wmflabs.org/v4/marker/pin-l-theatre+000@2x.png

"cover" makes sure this relative scaling is respected while still
giving us exactly the "clip at the bottom" effect we want.

Bug: T326530
Change-Id: Ic855de36701605a4bfffe085076f0118587d0569