Don't renable draggable when map is static
Was missed while doing I548806fedd0fe98155c413d1731697e5352c9344.

When a few lines above during the setup the initial static mode is
set. See toggleStaticState(). If the map is static, the drag and
zoom handlers are disabled for a reason. They should not be enabled
just like that.

Bug: T318510
Change-Id: I688ed42bfa83f04580105777154db4c40165df89