Add strict types to properties and parameters where possible
It's not always possible without also changing the code. But this
patch is intentionally not touching anything else.

This also fixes a few minor mistakes that surfaced when we tried to
add the corresponding strict types.

Minor updates before large refactoring for Parsoid compatibility. This
makes it easier to track relevant IDE warnings later down the line.

Bug: T263762
Change-Id: I1d5fb05350da6739b22c5b92155b1a8570a4a1ab