Fix and add mising parser test for maplink with suppressed text=""
This broke with I0a2c4d9 because it was never tested and the old
code was pretty much unreadable. The meaning of the old line

 $this->getText( 'text', null, '/\S+/' )

was actually this:
* Because of the `null` the parameter is optional and would default
  to `null`, triggering the following `if`.
* Non-empty text="…" do what you expect.
* An empty text="" will be turned into false. This is passed to
  recursiveTagParse, which is not even allowed and will be reported
  in later PHP 8 versions, just to be turned back into an empty
  string.

This back and forth turns out to be just pointless. Just keep the
empty string.

It's not entirely clear if this was ever an intended feature.
But because it's used on live wikis we must respect it as if it's a
feature.

Bug: T328739
Change-Id: I502b23963a96d0f07434adf67fa38217b79f9ffd