Fix incomplete parse-time expansion for "page" data source
Must do the same as here:
https://github.com/wikimedia/mapdata/blob/master/src/ExternalDataParser.js#L48
Appears like we missed this entirely before. Luckily this is not a
huge issue because the client-side code is still in place and would
successfully resolve this just as before. This specific parse-time
expansion was just pointless. Just made the HTML bigger without any
other effect.

Bug: T323113
Change-Id: Id403ca9cffa2bc5c96081d3482d723f38402829d