Make enter work on fullscreen map close button
This is the only button that is created in our code. The other
buttons (zoom in/out as well as the layer button below the close
button) are created by the mapbox lib. The only difference I can
find is the href="#". This is added to all buttons except the close
button. When I add it pressing enter on the button starts working.
It appears like this is reqired so the browser converts the enter
press into a click event.

In HTML5 an <a> tag without a href="…" is not a link but simple
visual inline formatting (like <b>) without any interaction.

Bug: T326804
Change-Id: I7c642a77e7bdcf116875730df80e88f73c887ae6