Don't tab to non-functional fullscreen button in static mode
Turns out there is no need to tab to this button. Tabbing to the
static <mapframe> and pressing enter there is enough.

We might need to revisit this UX-wise. E.g. it might be better to
tab to the button but not to the map. But this would need much more
code. I suggest to start merging this as a trivial 1st step and ask
this question in demo time.

Bug: T326786
Change-Id: Ied7c7d38e657cadfbd5a60fd3e4d4e847a1e2a0c