Nearby: Fix bogus popup size calculation
Problem: Because of the way `max-width: 100%` works the popup size is
calculated only based on the text in the popup, completely ignoring
the image. This results in very inconsistent image sizes, often way
smaller than 250px in popups that don't have much text.

Turns out we can use an absolute value like 250px instead, no problem.
The total popup width is limited to 300px, hard-coded. Subtracting
margins and paddings leaves about 276px for the content in the popup.
We can round this down to one of the standard thumbnail sizes: 250px.

Unfortunately the pageimages API acts inconsistent on upright
images. To avoid having extremely tall popups we add a max-height as
well.

Note this CSS is for both nearby popups as well as GeoJSON objects
with a description. (Also for the old nearby module, but that should
be mostly irrelevant by now.) User-provided GeoJSON can contain
images in any size. The main purpose of this CSS is to limit such
user-provided images.

While we are here we fix the minWidth of the nearby popups to be
identical to the popups that come from the GeoJSON. This shouldn't
have much of an effect any more with the change above. Only in
popups with no image and very short text.

Bug: T327926
Change-Id: Id085036edf4ba88ff3b339677391c601c3aa7cda