Simplify custom pointToLayer callback to use upstream code
I experimented with something similar in I83d821d. Seems to be an
almost perfect fit here. This just calls the original function and
does nothing but change an option.

However. It might be that this entire pointToLayer hack becomes
pointless with I1fbd9fb in place. It should be removed then.

Bug: T326790
Change-Id: I69e7ec42d0c4791a20553197299d319ceb299a3a