Don't create impossible Special:Map/a/a/a links
The special page doesn't even understand this. Note that the (float)
casts have been added just today via I154322b. That patch already
changed Special:Map/a/a/a to Special:Map/0/0/0. I think this was a
mistake. It makes the link appear less broken but really just hides
an error and is more confusing because of that. We know there are no
coordinates and shouldn't make up some.

Note this was never tested.

Note the "a" was never documented as a possible input for the link.

The <a> element is still created because other processes might need
it, e.g. gadgets.

Bug: T223880
Change-Id: I4416230e047d172022c6193b94b16115920708d5