Hide the center of a marker cluster when expanded
This was a requirement in T308227, as far as I can tell: "We want
to hide the cluster marker when we show the spider. ( currently
it's blurred )". From a UX perspective it does indeed feel much
better than any other solution we tried so far (e.g. making it
partly non-interactive while still showing it, see old patchsets).

Also increase the z-index. This is relevant for the expanded points
in the cluster. When there are many points close to each other the
blue focus rectangle appears behind them. This is confusing. It
should always be in the very front.

Note: Having an element focused is a exceptional event. Using a
z-index that high shouldn't be a problem in this special case. It's
unlikely there is ever anything more important than the currently
focused element in the same spot.

Bug: T326790
Bug: T327159
Change-Id: I7557b377b06d201fd07fa7dad4adf6285af0fad1