Final VE fixes for deprecated (but supported) width in percent
Follow-up for the changes started in Iafe3866. Percentages other than
100% are still accepted in the wikitext by the backend code, but act
as if width="300" was entered.

In this particular case we intentionally cause a "dirty" diff by
replacing the discontinued percentage with the actual value. This also
helps keeping the code simple.

Additionally it was still possible to crash the dialog by entering
non-numeric gibberish into the width/height fields.

As before the goal of these patches is not to find the best possible
UX, but only:
1. Don't let the UI crash just because the user uses the keyboard.
2. Don't write gibberish to the wikitext.

Bug: T328541
Change-Id: I185cc6191ab908246a27b7c865b62a707c0e25e2