Fix fullscreen map sidebar not staying collapsed
This is clearly a mistake. Note the comment. Maybe the idea was to
actually destroy the sidebar when it's closed. But this is not what
happens. Initially it doesn't exist and this code works. But after
the sidebar was opened once it's never destroyed again, only hidden.

The toggle buttons state is where the information is stored if the
sidebar should be open or not.

Also replace direct access to the .value property with OOUI's
official .getValue().

Bug: T328545
Change-Id: Ib583a4247d9fd9dc5222aad9d8feadb26ec82f12