Drop duplicate tracking of valid <map…> tags
Tracking valid tags via State::setValidTags() is the same as tracking
the total number of usages via State::incrementUsage() minus the
number of invalid tags. This became more clear now that we started
separating the code much more.

Notes:
* We can drop the extra `if ( $count )` check. The same filtering
  already happens in the State class.
* This patch changes the JSON serialization. A bool becomes an int.
  This can't affect the "broken" tracking because true is 1 and 1 is
  tracked as broken, as before. But it can affect the "valid" tracking.
  In some cases a page that contains multiple broken tags but no valid
  tag might briefly appear in the valid category. I think this is
  acceptable.

I believe this change is worth it just because it gets rid of multiple
weird state changes in weird places.

Change-Id: I2435b22fd3acde7de7dada86fcccb79f36a83707