Use OOUI getHoldProcess to disable dragging as early as possible
See T297519 for a longer explanation. It might turn out that this
patch is not enough to fully solve the problem. For example, there
are more interactions that might need to be disabled the same time
(namely doubleClickZoom, scrollWheelZoom, and keyboard). Still I
suggest to start simple and continue monitoring our error logs.

Bug: T297519
Change-Id: I34dd8c0376990b0881bd3539500fa03fb373b902