Extract separate little ErrorReporter service
The main benefit is that this removes a good chunk of really
isolated code from the abstract base class. This makes it easier
to test, and more obvious what the (remaining) purpose of the base
class is. Long-term goal is to possibly get rid of the base class.

Re-use is not a concern as this code can't be reused for the
Parsoid port anyway.

Change-Id: I58ec0ee28b017007ac74098b1c0225f6effecd34