Make updateMarkerSymbolCounters data update optional
When processing Kartographer extension tags, the legacy parser keeps a
state that enables an immediate replacement of the counter types by
their instantiations.
In Parsoid, we want to not have a share state between extension tag
invocations, so we want to keep the type of counter in the data until
post-processing.
We split the "updateMarkerSymbolCounters" to create a version that is
not updating said counters but only returning the first counter of the
geometries.

Bug: T263762
Change-Id: Ifd501ac22d5b2c1ad659962f679ee4b986091e46