Replace MapLinkTrait by MapLinkAttributeGenerator
The call to the trait was pretty verbose and the resulting code depended
on correct associative array wrangling, which would have been repeated
in the ParsoidMapLink.

Hence, we pass a the newly created MapTagArgumentValidator instead.
Since we can't easily move this property to the maplink (and hence to
the trait), the trait feels harder to justify; so we transform it into a
regular class.

Bug: T263762
Change-Id: I8431285b832a25d34e91e35b235e27676caa036b