Introduce MapFrameAttributeGenerator
On the same model as MapLinkAttribute generator, we export attribute
generation to a MapFrameAttributeGenerator that can be used by both the
LegacyMapFrame and the ParsoidMapFrame.

The parserTests updates are attribute reordering that happen because a
the alt attribute's computation is language dependent and needs to be
handled differently in Parsoid; it's consequently left in the
LegacyMapFrame class.

Bug: T263762
Change-Id: I2916b9cfc5eef818b8f83f58bab1097652a715f2