Add a Parsoid skeleton
This is a "do nothing" patch that should normally have no impact on the
current parsing. We're only adding the Parsoid infrastructure that will
eventually be used to implement the tags. Since no method is
implemented, everything is pass-through and the legacy parser is still
used as a fallback here.

The Parsoid parserTests are introduced to have a reference point for
what we're doing/modifying. It currently shows what is done with the
legacy fallback, and have no guarantee of correctness.

Bug: T263762
Change-Id: Ia362dc52823aa625de89f1080f3bf0614a054ae1