Inline all $static… MapFrameAttributeGenerator properties
Turns out the three "a" for auto-zoom and auto-position are only
really needed in a single place. This patch moves this knowledge to
this one spot in the new MapFrameAttributeGenerator class. We don't
need to bother SpecialMap or anybody else with this. These consumers
get null now.

This also fixes an inconsistency between <mapframe> and <maplink>.
Since I4416230 we don't create broken, dysfunctional links to
[[Special:Map/0/a/a]] any more. With this patch the behavior is now
the same for <maplink>s.

Bug: T263762
Change-Id: Icc4d142f51a27094844ca2993db4adf0d1f18a75