Remove now unused expensive parser function instrumentation
This is unused after I323c744. It's easy to add back in case we need
it again. It was a little awkwardly placed anyway because it meant
there are two almost entirely unrelated features in the same
interface.

I'm also moving the file out of the Tag namespace up to the other
small services we extracted from the actual tag handling. We need
to reorganize this codebase further, e.g. separate Core Parser and
Parsoid better, but this is for later.

Bug: T332973
Change-Id: I8b89000905048c1d00812fa9d5c2ccdd6c8a6f0a