styles: Flag wrongly applied token
The Codex design tokens delivered via 'mediawiki.skin.variables' follow
a naming scheme in order to not apply tokens to the wrong elements
properties.
`color-` tokens are only applied to `color` properties for foreground
text color. Compare the available border tokens at
https://doc.wikimedia.org/codex/latest/design-tokens/border.html
All of this is to support multi-themeability and best-possible
consistency.

Change-Id: Iefe8d3c8618713f2ef4631587846a5ae3118f572