Reduce line-height for long fulscreen map captions
These greatly benefit from a slightly more narrow line height.

The original line-height was a hard-coded 31px, added in 2016 via
I6224ffb. It's really hard to tell what the reasoning was. The 1.55
you see in the current code are a 1:1 mapping of these 31px to a
value that behaves relative to the font size, while still resulting
in 31px. See Iededaf6.

Bug: T335032
Change-Id: Id717daf91ac2584b0e8e3db4851ba8134da56944