Streamline the fullscreen footer setup
There's already a guard checking for if the dialog is opened for
the first time, there should be no need to check each element
again.

There's also no sense in setting the caption text in this step. The
map will always be set by a call to setMap() that's done after the
setup. There the caption text will be applied.

Also re-ordering the setup to follow the order in the DOM.

Change-Id: I81229e0f67783da137b16bc0016a150bafddf162